Added a possibility to provide a package blacklist
If a package is causing problems the only way to mitigate this is to
uninstall the package before building the bridge and re-installing it
afterwards. This may however not be something that easy to do, as such
packages could potentially have a large dependency tree following them.

This commits adds the possibility to provide a 'blacklist.txt' file
listing packages that should simply be ignored by the bridge.